+++ modules/hanna-react/CHANGELOG.md
@@ -10,6 +10,11 @@
   - feat: Add prop `renderToggler()`, deprecate `Toggler` (comp) prop instead
 - `Pagination`:
   - fix: Suppress annoying dev-mode warning message about `key` prop access
+- `Skeleton`:
+  - feat: Add prop `circle`
+  - feat: Add `Skeleton.circle` helper, and branded `SkeletonCircle` type
+  - feat: Emit console warnings when unsupported prop combinations are used
+  - fix: `wrapperElmProps` are now actually applied
 
 ## 0.10.173